Refactor to be more bdd and follow clean coding approaches
