+++ src/components/editor/page/download.no-pagelayouts.test.tsx
@@ -2,7 +2,7 @@ import { describe, it, expect, vi } from 'vitest';
 import { render} from '@testing-library/react';
 import { Download } from './download';
 import { Editor } from '@craftjs/core';
-import * as LocalData from '../local-data';
+import * as LocalData from '../page-layout';
 
 describe('Given the download component', async () => {
   //Arrange
